backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
requests==2.31.0
psycopg2-binary==2.9.9
